refactor(api): neutralize agentphone and email service filenames
Rename the AgentPhone and shared-email service modules to domain-neutral
filenames and update every referencing module. Every export in both files
was already neutral, so the moves are pure 100% renames with no body
changes.

- services/zero-agentphone.service.ts -> services/agentphone.service.ts
- services/zero-email-common.service.ts -> services/email-common.service.ts

Eleven import specifiers move to the new paths. The agent compose
consolidation preflight manifest pins scanned repository paths, so its
three entries for the agentphone service move with them; the entries are
otherwise unchanged and stay in collector sort order.

Closes #27917
