+++ turbo/apps/api/src/signals/services/email-common.service.ts
