+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -174,6 +174,8 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/agent-runs.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/agentphone-shared.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/agentphone-shared.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
+    "turbo/apps/api/src/signals/services/agentphone.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/agentphone.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/chat-goal-queue.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/chat-search.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/chat-thread-event.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -217,8 +219,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/webhooks-clerk-cleanup.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-automation.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-agentphone.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
-    "turbo/apps/api/src/signals/services/zero-agentphone.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-browser.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
@@ -282,6 +282,7 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/agentphone-chat-ingress.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/agentphone-queued-launch-context.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/agentphone-shared.service.ts|agentComposes,zeroAgents",
+    "turbo/apps/api/src/signals/services/agentphone.service.ts|agentComposeId,agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/banking.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/browser-authorization.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/canonical-feishu-ingress-processor.service.ts|agentComposeId",
@@ -347,7 +348,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/workflow-data.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-user-automation-thread.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/workflow-volume.service.ts|headVersionId",
-    "turbo/apps/api/src/signals/services/zero-agentphone.service.ts|agentComposeId,agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-browser.service.ts|agentComposeId,agentComposes",
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-chat-title.service.ts|agentComposeId",
